# docs/api-contract.openapi.yaml
openapi: "3.1.0"
info:
  title: "Club Informatique de l’IST : contrat d’API"
  version: "1.0.0"
  description: |
    Contrat de référence entre le frontend et le backend.

    Conventions : dates en ISO 8601 UTC ; pagination uniforme (« content », « page », « size », « totalElements », « totalPages ») ;
    erreurs au format RFC 9457 avec un code stable et des messages en français ; identifiants numériques stables.
    Le jeton d’accès est transmis par l’en-tête « Authorization » ; le jeton de rafraîchissement par un cookie HttpOnly.
    L’extension « x-roles » liste les rôles autorisés ; un Formateur et un Responsable du Club héritent des droits du Membre, un Super Admin de ceux de l’Administrateur.
    Les droits sont appliqués par le serveur, par rôle et par propriété de la ressource.
servers:
  - url: /api/v1
...
security:
  - jetonAcces: []
paths:
...
  "/evenements":
    get:
      tags: [Contenus publics]
      summary: Événements publiés.
      operationId: getEvenements
      security: []
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/categorieId"
        - name: aVenir
          in: query
          schema:
            type: boolean
          description: Ne retenir que les événements non terminés.
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageEvenement"
        "400":
          $ref: "#/components/responses/Erreur400"
    post:
      tags: [Gestion du club]
      summary: Créer un événement.
      operationId: postEvenements
      x-roles: [RESPONSABLE_CLUB]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/EvenementSaisie"
      responses:
        "201":
          description: Ressource créée.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Evenement"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/evenements/slug/{slug}":
    get:
      tags: [Contenus publics]
      summary: Événement publié.
      operationId: getEvenementsSlugSlug
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Evenement"
        "404":
          $ref: "#/components/responses/Erreur404"
  "/evenements/{id}/calendrier":
    get:
      tags: [Contenus publics]
      summary: Événement au format iCalendar (décision D-06).
      operationId: getEvenementsIdCalendrier
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Fichier iCalendar.
          content:
            "text/calendar":
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Erreur404"
...
  "/gestion/evenements":
    get:
      tags: [Gestion du club]
      summary: Tous les événements, publiés ou non.
      operationId: getGestionEvenements
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: du
          in: query
          schema:
            type: string
            format: date
          description: Début de la période (date de début de l’événement).
        - name: au
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/PageEvenement"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
  "/evenements/{id}":
    put:
      tags: [Gestion du club]
      summary: Modifier un événement. 409 si la capacité devient inférieure au nombre d’inscrits ; une capacité accrue profite à la liste d’attente.
      operationId: putEvenementsId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/EvenementSaisie"
      responses:
        "200":
          description: Succès.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Evenement"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
        "409":
          $ref: "#/components/responses/Erreur409"
    delete:
      tags: [Gestion du club]
      summary: Supprimer un événement ; ses inscrits sont désinscrits et prévenus.
      operationId: deleteEvenementsId
      x-roles: [RESPONSABLE_CLUB]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Opération réalisée, sans contenu.
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
...
  "/fichiers":
    post:
      tags: [Fichiers]
      summary: Déposer un fichier (taille, type et signature binaire contrôlés).
      operationId: postFichiers
      x-roles: [FORMATEUR, RESPONSABLE_CLUB]
      requestBody:
        required: true
        content:
          "multipart/form-data":
            schema:
              type: object
              required: [fichier]
              properties:
                fichier:
                  type: string
                  format: binary
      responses:
        "201":
          description: Ressource créée.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Fichier"
        "400":
          $ref: "#/components/responses/Erreur400"
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "413":
          $ref: "#/components/responses/Erreur413"
        "415":
          $ref: "#/components/responses/Erreur415"
  "/fichiers/{id}":
    get:
      tags: [Fichiers]
      summary: Télécharger un fichier, selon les droits de la ressource qui le porte.
      operationId: getFichiersId
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contenu du fichier.
          content:
            "application/octet-stream":
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Erreur401"
        "403":
          $ref: "#/components/responses/Erreur403"
        "404":
          $ref: "#/components/responses/Erreur404"
components:
  securitySchemes:
    jetonAcces:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Numéro de page, à partir de zéro.
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
      description: Taille de page.
    sort:
      name: sort
      in: query
      schema:
        type: string
      description: "Tri : « champ,asc » ou « champ,desc »."
    search:
      name: search
      in: query
      schema:
        type: string
      description: Recherche plein texte ; facultative.
    categorieId:
      name: categorieId
      in: query
      schema:
        type: integer
        format: int64
  responses:
    Erreur400:
      description: "Requête invalide : erreurs de champ dans « errors »."
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur401:
      description: Session absente ou expirée, ou identifiants refusés.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur403:
      description: Droits insuffisants (rôle ou propriété de la ressource).
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur404:
      description: Ressource introuvable.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur409:
      description: Conflit avec l’état actuel des données.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur413:
      description: Fichier trop volumineux.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
    Erreur415:
      description: Type de fichier refusé.
      content:
        "application/problem+json":
          schema:
            $ref: "#/components/schemas/Probleme"
...
  schemas:
    Probleme:
      type: object
      description: Erreur au format RFC 9457. Le champ « code » est stable ; « detail » est rédigé en français.
      required: [title, status, code]
      properties:
        type:
          type: [string, "null"]
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int64
        detail:
          type: [string, "null"]
        instance:
          type: [string, "null"]
        code:
          type: string
        errors:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/ErreurChamp"
    ErreurChamp:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    Evenement:
      type: object
      description: « nombreInscrits » et « placesRestantes » sont calculés en base sur les inscriptions confirmées.
      required: [id, titre, slug, description, dateDebut, dateFin, lieu, nombreInscrits, publie]
      properties:
        id:
          type: integer
          format: int64
        titre:
          type: string
        slug:
          type: string
        description:
          type: string
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        lieu:
          type: string
        capaciteMax:
          type: [integer, "null"]
          format: int64
        nombreInscrits:
          type: integer
          format: int64
        placesRestantes:
          type: [integer, "null"]
          format: int64
        publie:
          type: boolean
        categorieId:
          type: [integer, "null"]
          format: int64
        categorieNom:
          type: [string, "null"]
        organisateurNom:
          type: [string, "null"]
    EvenementSaisie:
      type: object
      required: [titre, description, dateDebut, dateFin, lieu, publie]
      properties:
        titre:
          type: string
        description:
          type: string
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        lieu:
          type: string
        capaciteMax:
          type: [integer, "null"]
          format: int64
        categorieId:
          type: [integer, "null"]
          format: int64
        publie:
          type: boolean
...
    Fichier:
      type: object
      required: [id, url, nom, type, tailleOctets]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        nom:
          type: string
        type:
          type: string
        tailleOctets:
          type: integer
          format: int64
...
    PageEvenement:
      type: object
      description: Page de résultats (pagination uniforme).
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Evenement"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...